config/urls.py
from django.contrib import admin
from django.conf.urls.static import static
from django.conf import settings
from django.urls import path, include
from rest_framework.schemas import get_schema_view
from django.views.generic import TemplateView


urlpatterns = [
    path('', TemplateView.as_view(template_name="products/swagger-ui.html")),
    path('admin/', admin.site.urls),
    path('products/', include('products.urls')),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
] + static(settings.STATIC_URL)
