requirements.txt
Django==3.2.4
django-extensions==3.1.3
djangorestframework==3.12.4
pyOpenSSL==20.0.1
pytest==6.2.4
pytest-django==4.4.0
PyYAML==5.4.1
uritemplate==3.0.1
Werkzeug==2.0.1
